Moved js and py solutions per directories.
Extracted ogg opus parser into separate files.
