+++ examples/js/package.json
@@ -1,0 +1,7 @@
+{
+    "dependencies": {
+        "configparser": "0.3.6",
+        "ws": "7.3.1",
+        "buffer": "5.6.0"
+    }
+}
